Move OTP implementations into class; update docs (#12)
This is mostly an internal structural refactor; there should be no breaking changes. This does adjust the internals of TOTP's counter calculation to use the system clock instead of the server request time (which was originally motivated by a mix of making tests work and a micro-optimization picked up from Rasmus years ago that could actually lead to issues in long-running processes occasionally used in modern deployments).

The public API expands to have an OO interface, which is wrapped by the original function-based versions.